# specs/001-todo-security-hardening/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo Security Hardening API
  version: 1.0.0
  description: API contract for authenticated todo management with user ownership isolation.
servers:
  - url: https://api.example.com
security:
  - bearerAuth: []
paths:
  /auth/signup:
    post:
      summary: Register a new user account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: Account created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '409':
          description: Username already exists
  /auth/login:
    post:
      summary: Authenticate user and issue session token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
  /todos:
    get:
      summary: List authenticated user's todos
      responses:
        '200':
          description: User-owned todo list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
        '401':
          description: Missing or invalid token
    post:
      summary: Create a todo for authenticated user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoCreate'
      responses:
        '201':
          description: Todo created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '401':
          description: Missing or invalid token
  /todos/{todoId}:
    patch:
      summary: Update completion state of authenticated user's todo
      parameters:
        - name: todoId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoUpdate'
      responses:
        '200':
          description: Todo updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '401':
          description: Missing or invalid token
        '403':
          description: Attempted cross-user access
        '404':
          description: Todo not found
    delete:
      summary: Delete authenticated user's todo
      parameters:
        - name: todoId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Todo deleted
        '401':
          description: Missing or invalid token
        '403':
          description: Attempted cross-user access
        '404':
          description: Todo not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SignupRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 8
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [access_token, token_type, user]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: bearer
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      required: [id, username]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
    Todo:
      type: object
      required: [id, task, is_done, user_id]
      properties:
        id:
          type: string
        task:
          type: string
        is_done:
          type: boolean
        user_id:
          type: string
          format: uuid
    TodoCreate:
      type: object
      required: [task]
      properties:
        task:
          type: string
          minLength: 1
    TodoUpdate:
      type: object
      required: [is_done]
      properties:
        is_done:
          type: boolean